+++ trunk/src/libgpuarch/cf-engine.c
@@ -113,7 +113,7 @@ void gpu_cf_engine_fetch(struct gpu_compute_unit_t *compute_unit)
 			compute_unit->id,
 			uop->work_group->id,
 			wavefront->id,
-			(long long) uop->id_in_compute_unit,
+			uop->id_in_compute_unit,
 			str2);
 	}
 }
@@ -154,7 +154,7 @@ void gpu_cf_engine_decode(struct gpu_compute_unit_t *compute_unit)
 		"cu=%d "
 		"uop=%lld\n",
 		compute_unit->id,
-		(long long) uop->id_in_compute_unit);
+		uop->id_in_compute_unit);
 }
 
 
@@ -240,7 +240,7 @@ void gpu_cf_engine_execute(struct gpu_compute_unit_t *compute_unit)
 		"cu=%d "
 		"uop=%lld\n",
 		compute_unit->id,
-		(long long) uop->id_in_compute_unit);
+		uop->id_in_compute_unit);
 }
 
 
@@ -283,7 +283,7 @@ void gpu_cf_engine_complete(struct gpu_compute_unit_t *compute_unit)
 			"cu=%d "
 			"uop=%lld\n",
 			compute_unit->id,
-			(long long) uop->id_in_compute_unit);
+			uop->id_in_compute_unit);
 		if (debug_status(gpu_stack_debug_category))
 			gpu_uop_debug_active_mask(uop);
